moved components folder outside of Pages, changed pathing in pages to reflect.
