+++ src/pages/about.js
@@ -1,7 +1,7 @@
 // about.js
 
 import React from 'react'
-import Layout from './components/layout'
+import Layout from '../components/layout'
 
 const About = () => {
   return (
