+++ src/pages/contact.js
@@ -1,7 +1,7 @@
 // contact.js
 
 import React from 'react'
-import Layout from './components/layout'
+import Layout from '../components/layout'
 
 const Contact = () => {
   return (
