+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import * as React from "react"
-import Layout from './components/layout'
-import CallToAction from './components/callToAction'
-import Bio from './components/bio'
+import Layout from '../components/layout'
+import CallToAction from '../components/callToAction'
+import Bio from "../components/bio"
 
 // data
 
@@ -15,7 +15,7 @@ const IndexPage = () => {
 
         <CallToAction>This is a Call To Action</CallToAction>
         
-        <Bio showCount={20} showStatus={false}>
+        <Bio showCount={20}>
 
           This is my Bio, it will be an expandable bio, but right not it is just a simple div.
